Put all CC helper classes into helper_cc.py
